[Refactor][Xiaomeng] add pages,routes folders under the src.
