+++ frontend/src/pages/config/PageConfig.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
 import { Layout, Steps, Typography } from "antd";
-import { ConfigStep } from "../shared/__types__/base";
+import { ConfigStep } from "../../shared/__types__/base";
 import ConfigSuccess from "./components/ConfigSuccess";
 import ProjectNameSetup from "./components/ProjectNameSetup";
-import PipelineSetup, { FormValues } from "../shared/components/PipelineSetup/PipelineSetup";
-import { BaseProject, createProjectUsingPost } from "../shared/clients/projectApis";
+import PipelineSetup, { FormValues } from "../../shared/components/PipelineSetup/PipelineSetup";
+import { BaseProject, createProjectUsingPost } from "../../shared/clients/projectApis";
 
 const { Paragraph } = Typography;
 const { Step } = Steps;
