+++ frontend/src/pages/dashboard/components/MetricInfo.tsx
@@ -1,10 +1,10 @@
 import React, { FC } from "react";
-import { ColourLegend } from "../../shared/components/ColourLegend";
-import { metricsStanderMapping, metricsExplanations } from "../../shared/constants/metrics";
+import { ColourLegend } from "../../../shared/components/ColourLegend";
+import { metricsStanderMapping, metricsExplanations } from "../../../shared/constants/metrics";
 import { Typography } from "antd";
-import { GRAY_1 } from "../../shared/constants/styles";
-import { MetricType } from "../../shared/__types__/base";
-import { MetricsUnit } from "../../shared/clients/metricsApis";
+import { GRAY_1 } from "../../../shared/constants/styles";
+import { MetricType } from "../../../shared/__types__/base";
+import { MetricsUnit } from "../../../shared/clients/metricsApis";
 
 const { Title, Paragraph } = Typography;
