+++ frontend/src/routes/Routes.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import { PageDashboard } from "./dashboard/PageDashboard";
-import { PageConfig } from "./config/PageConfig";
+import { PageDashboard } from "../pages/dashboard/PageDashboard";
+import { PageConfig } from "../pages/config/PageConfig";
 import { Redirect, Route, Switch } from "react-router-dom";
 import { renderRoutes, RouteConfig } from "react-router-config";
-import { useQuery } from "./shared/hooks/useQuery";
+import { useQuery } from "../shared/hooks/useQuery";
 
 const INITIAL_ROUTES: RouteConfig[] = [
 	{
